PeriodeOgLand -> Periode
