+++ src/felleskomponenter/menypanel/menypunkter/periode/index.ts
@@ -1,0 +1,3 @@
+import Periode from './periode';
+
+export default Periode;
