+++ src/felleskomponenter/menypanel/menypunkter/periode/periode.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+
+import * as Nav from '../../../../utils/navFrontend';
+import * as Etiketter from '../etiketter';
+import * as KV from '../../../../kodeverk';
+
+import Soknadsperiode from './soknadsperiode';
+
+interface PeriodeOgLandProps {
+  visArbeidsforholdRolleEtiketter: boolean,
+  redigerbart: boolean,
+  lagreSoknadOgOppfriskSaksopplysninger: () => void,
+}
+
+const PeriodeOgLand = ({
+  visArbeidsforholdRolleEtiketter,
+  redigerbart,
+  lagreSoknadOgOppfriskSaksopplysninger,
+}: PeriodeOgLandProps) => (
+  <div>
+    <div style={{ marginBottom: '1em' }}>
+      <Nav.typo.Undertittel style={{ display: 'inline', marginRight: '1em' }}>{KV.Menypunkter.Periode.tittel}</Nav.typo.Undertittel>
+      <Etiketter.FraSoknad style={{ marginRight: '0.3em' }} />
+      {
+        visArbeidsforholdRolleEtiketter &&
+        <Etiketter.ArbeidsgiversDel />
+      }
+    </div>
+    <Nav.Row>
+      <Nav.Column xs="6">
+        <Soknadsperiode
+          redigerbart={redigerbart}
+          lagreSoknadOgOppfriskSaksopplysninger={lagreSoknadOgOppfriskSaksopplysninger}
+        />
+      </Nav.Column>
+    </Nav.Row>
+  </div>
+);
+
+export default PeriodeOgLand;
