+++ src/felleskomponenter/menypanel/menypunkter/periode/soknadsperiode/index.js
