+++ src/felleskomponenter/menypanel/menypunkter/periode/soknadsperiode/soknadsperiode.js
