+++ src/felleskomponenter/menypanel/menypunkter/periode/soknadsperiode/soknadsperiode.less
