+++ src/felleskomponenter/menypanel/menypunkter/periode/soknadsperiode/soknadsperiode.test.js
