--- src/felleskomponenter/menypanel/menypunkter/periodeOgLand/index.ts
@@ -1,3 +1,0 @@
-import PeriodeOgLand from './periodeOgLand';
-
-export default PeriodeOgLand;
