+++ src/felleskomponenter/menypanelForm/soknad.tsx
@@ -132,7 +132,7 @@ const Soknad = ({
           'Arbeidsforhold og inntekt',
           'Arbeidsgiver/virksomhet',
           'Fullmektig',
-          'Periode og land',
+          'Periode',
           'Utenlandsoppdraget',
           'Lønn og godtgjørelser',
           'Arbeidssted(er)',
